리뷰등록, 사용자 리뷰 출력 완료

+++ app/src/main/java/kr/racto/milkyway/model/RoomData.kt
@@ -4,6 +4,6 @@ data class RoomData(
     val reviewList: List<Review>,
     val roomName: String,
     val address: String,
-    val ratingAvg: Int,
+    val ratingAvg: Double,
     val reviewCount: Int
 )
